requirements.txt
boto3==1.34.148
botocore==1.34.148
jmespath==1.0.1
numpy==2.0.1
pandas==2.2.2
pdf2image==1.17.0
pillow==10.4.0
psycopg2-binary==2.9.9
python-dateutil==2.9.0.post0
python-magic==0.4.27
pytz==2024.1
s3transfer==0.10.2
six==1.16.0
tzdata==2024.1
urllib3==2.2.2
